local test enhance

+++ functions/data-processing-engines/simple-dataform-query-executor/test.sh
@@ -1,12 +1,28 @@
+#!/bin/bash
+
+delete_dataform_repo=true
+if [ $# -gt 0 ]; then
+    delete_dataform_repo=$1
+fi
+
 # Project and environment variables
 project=pso-amex-data-platform
 location=us-central1
 repository_id=test-repo5
 
 # SQL query and owner information
+dataform_location=us-central1
+dataform_project_id=pso-amex-data-platform
 commitname=bqfile
-filename=querytest.sql
+
+definitions_dir=definitions
+workflow_name=workflow1
+job_name=J01_etl_step_1
+
+filepath="${definitions_dir}/${workflow_name}/${job_name}.sqlx"
+
 query='SELECT * FROM `bigquery-public-data.austin_crime.crime` where clearance_date>${dataform.projectConfig.vars.start_date} LIMIT 1000'
+start_date="2019-01-01"
 queryowner=oscarpulido@google.com
 query=$(echo "$query" | tr -d '\n')
 encoded_query=$(echo -n "$query" | base64)
@@ -41,7 +57,7 @@ curl -X POST \
             "commitMessage": "update bq query"
           },
           "fileOperations": {
-            "'$filename'": {
+            "'$filepath'": {
               "writeFile": {
                 "contents": "'$encoded_query'"
               }
@@ -63,26 +79,27 @@ curl -X GET \
 # Read the contents of the committed SQL file
 curl -X GET \
      -H "Authorization: Bearer $(gcloud auth print-access-token)" \
-     https://dataform.googleapis.com/v1beta1/projects/$project/locations/$location/repositories/$repository_id:readFile?path=$filename
+     https://dataform.googleapis.com/v1beta1/projects/$project/locations/$location/repositories/$repository_id:readFile?path=$filepath
 
-# -----------------------------------------------------
-# Execute the SQL query via Cloud Function
-# -----------------------------------------------------
-curl -m 70 -X POST https://us-central1-pso-amex-data-platform.cloudfunctions.net/simple-dataform-query-executor \
--H "Authorization: bearer $(gcloud auth print-identity-token)" \
--H "Content-Type: application/json" \
--d '{
-  "location": "'$location'",
-  "dataform_project_id": "'$project'",
-  "bq_project_id": "'$project'",
-  "repository_name": "'$repository_id'",
-  "file_path":  "'$filename'",
-  "query_variables":{
-      "${dataform.projectConfig.vars.start_date}":"2019-01-01"
-  }
-}'
 
-# Delete the Dataform workspace
-curl -X DELETE \
-     -H "Authorization: Bearer $(gcloud auth print-access-token)" \
-     https://dataform.googleapis.com/v1beta1/projects/$project/locations/$location/repositories/$repository_id
+if [ "$delete_dataform_repo" = "true" ]; then
+    # -----------------------------------------------------
+    # Execute the SQL query via Cloud Function
+    # -----------------------------------------------------
+    curl -m 70 -X POST https://$location-$project.cloudfunctions.net/orch-framework-simple-dataform-query-executor \
+    -H "Authorization: bearer $(gcloud auth print-identity-token)" \
+    -H "Content-Type: application/json" \
+    -d '{
+      "dataform_location": "'$dataform_location'",
+      "dataform_project_id": "'$dataform_project_id'",
+      "repository_name": "'$repository_id'",
+      "file_path":  "'$filepath'",
+      "query_variables":{
+          "${dataform.projectConfig.vars.start_date}":"'$start_date'"
+      }
+    }'
+
+    curl -X DELETE \
+         -H "Authorization: Bearer $(gcloud auth print-access-token)" \
+         https://dataform.googleapis.com/v1beta1/projects/$project/locations/$location/repositories/$repository_id
+fi
